% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{addPosition}
\alias{addPosition}
\title{Add position columns to MODIS data retrieved from a SciDB's 3D array}
\usage{
addPosition(sdbdf, period, startyear)
}
\arguments{
\item{sdbdf}{A data frame made of MODIS data. The ID columns must be named as "col_id", "row_id", and "time_id"}

\item{period}{Number of days between observations (e.g 8)}

\item{startyear}{Initial year of the index (e.g 2000)}
}
\value{
A data frame with additional columns
}
\description{
Add position columns to MODIS data retrieved from a SciDB's 3D array
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
