% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{calcPixelSize}
\alias{calcPixelSize}
\title{Calculate the length of a MODIS pixel}
\usage{
calcPixelSize(resolution, tileWidth)
}
\arguments{
\item{resolution}{Square root of the number of pixels on an image}

\item{tileWidth}{Width of a tile}
}
\value{
A number
}
\description{
Calculate the length of a MODIS pixel. Resolution is the number of pixel in one dimension (e.g 4800)
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
