% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{calcTileWidth}
\alias{calcTileWidth}
\title{Calculate the width of a MODIS tile}
\usage{
calcTileWidth()
}
\value{
A number
}
\description{
Calculate the width of a MODIS tile
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
