% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{date2grid}
\alias{date2grid}
\title{year-day-of-the-year to time_id}
\usage{
date2grid(dateDOY, period, startyear)
}
\arguments{
\item{dateDOY}{Input day in year and day-of-the-year format (e.g 2001032 is Febraury the 2nd of 2001)}

\item{period}{Number of days between observations (e.g 8)}

\item{startyear}{Initial year of the index (e.g 2000)}
}
\value{
A number
}
\description{
DEPRECATED. Return a time index (timid) from the input date (MODIS DOY) and time period (e.g 8 days).
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
