% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getFilenameFromFilepath}
\alias{getFilenameFromFilepath}
\title{Get the filename from the path to the file}
\usage{
getFilenameFromFilepath(filepath)
}
\arguments{
\item{filepath}{Character representing the full path to the file}
}
\value{
Character representing the filename including the file extension
}
\description{
Return the filename of the path to the file
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
