% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getFilepathFromFilepath}
\alias{getFilepathFromFilepath}
\title{Get the filepath without the last part (filename)}
\usage{
getFilepathFromFilepath(filepath)
}
\arguments{
\item{filepath}{Character representing the full path to the file}
}
\value{
Character representing the filepath without the file name
}
\description{
Return the filepath of the path without the last part (filename)
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
