% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisutil.R
\name{getFirstGmip}
\alias{getFirstGmip}
\title{Return the GMPI of the first pixel (top left) of the given MODIS tile}
\usage{
getFirstGmip(modisTileId, nrows, ncols)
}
\arguments{
\item{modisTileId}{A character with a MODIS tile id (i.e "h10v08")}

\item{nrows}{Number of rows in a MODIS image (i.e for MOD09Q1 is 4800)}

\item{ncols}{Number of rows in a MODIS image (i.e for MOD09Q1 is 4800)}
}
\value{
Numeric vector containing the c(i,j) pixel coordinates in th GMPI
}
\description{
Return the GMPI of the first pixel (top left) of the given MODIS tile
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
