% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getHV}
\alias{getHV}
\title{Get tile's H and V from MODIS' tile Id}
\usage{
getHV(modisTileId)
}
\arguments{
\item{modisTileId}{A character with a MODIS tile id (i.e "h10v08")}
}
\value{
A character vector of 2 elements c(tH, tV)
}
\description{
Get the tileH and tileV from a MODIS tile Id
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
