% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getSdbDataFromBB}
\alias{getSdbDataFromBB}
\title{Get data over a SciDB connection of a bounding box}
\usage{
getSdbDataFromBB(con, arrayname, pixelSize, lonlat.mat, start, end, origin,
  period, yearly)
}
\arguments{
\item{con}{A SciDB connection object}

\item{arrayname}{A string. The name of the array}

\item{pixelSize}{A number. The length of one side of a pixel}

\item{lonlat.mat}{A 2x2 matrix. The 2 columns are the WGS84 longitude, and WGS84 latitude}

\item{start}{An integer. The start date as YYYYMMDD}

\item{end}{An integer. The end date as YYYYMMDD}

\item{origin}{An integer. A YYYYMMDD date. The day when the time_id == 0}

\item{period}{An integer. The number of days between observations}

\item{yearly}{A logical Do the dates yearly match January the 1st?}
}
\value{
A data frame
}
\description{
Get data over a SciDB connection of a bounding box
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
