% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getSdbDataFromPoints}
\alias{getSdbDataFromPoints}
\title{Get SciDB data from sample points}
\usage{
getSdbDataFromPoints(samples.df, lonlat, con, arrayname, pixelSize)
}
\arguments{
\item{samples.df}{A data.frame of samples. It must have 2 columns with the WGS84 longitude and latitude of the sample}

\item{lonlat}{A character vector. The names of the columns in samples.df that contain the  WGS84 longitude and latitude}

\item{con}{A SciDB connection object}

\item{arrayname}{A string. The name of the array}

\item{pixelSize}{A number. The length of one side of a pixel}
}
\value{
A list of lists. Each inner list has 2 elements: A row of samples.df and the results of the SciDB query for the longitude and latitude
}
\description{
Get data over a SciDB connection of a data.frame of sample points
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
