% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getTileIdFromFilename}
\alias{getTileIdFromFilename}
\title{Get the MODIS tile id from the modis filename}
\usage{
getTileIdFromFilename(fileName)
}
\arguments{
\item{fileName}{Name of the file}
}
\value{
The name of the file
}
\description{
Get the MODIS tile id from the modis filename
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
