% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisutil.R
\name{getTimeFromHdfFilename}
\alias{getTimeFromHdfFilename}
\title{Get the adquisition time of a MODIS HDF file name}
\usage{
getTimeFromHdfFilename(hdfFilename)
}
\arguments{
\item{hdfFilename}{HDF filename}
}
\value{
Character. A date in the format year and day of the year YYYYDOY
}
\description{
Get the adquisition time of a MODIS HDF file name
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
