% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisutil.R
\name{ids2tile}
\alias{ids2tile}
\title{Calculate the MODIS's tile index from the array's spatial indexes}
\usage{
ids2tile(col_id, row_id, nrows, ncols)
}
\arguments{
\item{col_id}{Array's col ID}

\item{row_id}{Array's row ID}

\item{nrows}{Number of rows in a MODIS image (i.e for MOD09Q1 is 4800)}

\item{ncols}{Number of rows in a MODIS image (i.e for MOD09Q1 is 4800)}
}
\value{
A list with 2 numeric values
}
\description{
Calculate the MODIS's tile index from the given array's spatial indexes
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
