% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{isLeapYear}
\alias{isLeapYear}
\title{Test for leap year}
\usage{
isLeapYear(year)
}
\arguments{
\item{year}{Numeric year}
}
\value{
TRUE is the year is leap, FALSE otherwise
}
\description{
Is the given year is a leap year?
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
