% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{missingtids}
\alias{missingtids}
\title{Report the missing time_ids}
\usage{
missingtids(tid)
}
\arguments{
\item{tid}{A vector of time ids}
}
\value{
A vecor with the missing time ids between the maximum and minimum time id provided
}
\description{
Report the missing time_ids
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
