% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{moveDateByYears}
\alias{moveDateByYears}
\title{Move dates certain number of years}
\usage{
moveDateByYears(Date1, numberOfYears)
}
\arguments{
\item{Date1}{A list made of Date objects}

\item{numberOfYears}{An integer number representing a of years}
}
\value{
Date1
}
\description{
Move a Date object a certain number of years, i.e 2000-10-31 moved 5 years becomes 2005-10-31
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
