% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{POSIXlt2txt}
\alias{POSIXlt2txt}
\title{Format a POSIXlt object}
\usage{
POSIXlt2txt(aPOSIXlt)
}
\arguments{
\item{aPOSIXlt}{A date object}
}
\value{
A string
}
\description{
Format a POSIXlt object
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
