% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{removeFisrtLast}
\alias{removeFisrtLast}
\title{Removes the first and last characters from a string}
\usage{
removeFisrtLast(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string
}
\description{
Removes the first and last characters from a single string
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
