% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{sampleTS}
\alias{sampleTS}
\title{Estimate the values for the time-series for the supplied sample times}
\usage{
sampleTS(ts.df, sampletime)
}
\arguments{
\item{ts.df}{data.frame with 2 columns: time and value}

\item{sampletime}{TODO}
}
\value{
a vector of sampled values
}
\description{
Estimate the values fo the time-series for the supplied sample times
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
