% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{scidbProcessSchema}
\alias{scidbProcessSchema}
\title{Parse a SciDB's array schema}
\usage{
scidbProcessSchema(schema)
}
\arguments{
\item{schema}{A string. Array schema - "<red:int16,nir:int16,quality:uint16> [col_id=48000:67199,1014,5,row_id=38400:52799,1014,5,time_id=0:9200,1,0]"}
}
\value{
A list containing 2 data.frame
}
\description{
Parse a SciDB's array schema
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
