% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisutil.R
\name{sinusoidal2gmpi}
\alias{sinusoidal2gmpi}
\title{Get the col_id & row_id in MODIS sinusoidal coordinates}
\usage{
sinusoidal2gmpi(lonlat.mat, pixelSize)
}
\arguments{
\item{lonlat.mat}{A numeric matrix with 2 columns: lon and lat on MODIS sinusoidal coordinates}

\item{pixelSize}{Pixel size in meters}
}
\value{
A 2-column matrix (col_id and row_id)
}
\description{
Calculate the col_id & row_id corresponding to the given MODIS sinusoidal coordinates
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
