% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{substrRight}
\alias{substrRight}
\title{Get the characters of a string from right to left}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{A string}

\item{n}{The number of chars}
}
\value{
The requested characters
}
\description{
Get the characters of a string from right to left
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
