% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{text2date}
\alias{text2date}
\title{Text to POSIXlt}
\usage{
text2date(dateAsText)
}
\arguments{
\item{dateAsText}{Date as string}
}
\value{
A date object (POSIXlt)
}
\description{
Transforms a text date given as text to a date object.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
