% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisutil.R
\name{time_id2date}
\alias{time_id2date}
\title{Transform a time_id into dates}
\usage{
time_id2date(time_id.vector, period)
}
\arguments{
\item{time_id.vector}{Vector of time indexes}

\item{period}{days between images (MOD09Q1 is 8, MOD13Q1 is 16)}
}
\value{
a list of Date objects
}
\description{
DEPRECATED. Transform a time_id into dates
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
