% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{ydoy2date}
\alias{ydoy2date}
\title{year-day_of_the_year to a date}
\usage{
ydoy2date(YYYYDOY)
}
\arguments{
\item{YYYYDOY}{Numeric or character with 4 digits for the year and 3 for the day of the year (i.e 2012324)}
}
\value{
A date object
}
\description{
DEPRECATED. Transform a date in the year-day_of_the_year format to a date
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
