% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{ydoy2ymd}
\alias{ydoy2ymd}
\title{Transform year-day-of-the-year into a date}
\usage{
ydoy2ymd(yyyydoy)
}
\arguments{
\item{yyyydoy}{An int YYYYDOY}
}
\value{
An integer YYYYMMDD
}
\description{
Transform year-day-of-the-year into a date
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
