% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisutil.R, R/time.R
\name{ymd2tid}
\alias{ymd2tid}
\alias{ymd2tid}
\title{Transform a date into a time_id index}
\usage{
ymd2tid(ymd, origin, period, yearly)

ymd2tid(ymd, origin, period, yearly)
}
\arguments{
\item{ymd}{An int YYYYMMDD}

\item{origin}{An int YYYYMMDD. The day when the time_id == 0}

\item{period}{An int. The number of days between observations}

\item{yearly}{A boolean. Do the dates yearly match January the 1st?}

\item{ymd}{An int. A YYYYMMDD date}

\item{origin}{An int. A YYYYMMDD date. The day when the time_id == 0}

\item{period}{An int. The number of days between observations}

\item{yearly}{A boolean. Do the dates yearly match January the 1st?}
}
\value{
An integer. The time_id matching ymd or 0 is ymd doesn't match

An integer. The time_id matching ymd or 0 (the first id) is ymd doesn't match
}
\description{
Calculate the width of a MODIS tile

Transform a date into an exactly time_id index
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}

Alber Sanchez, \email{alber.ipia@inpe.br}
}
