% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{ymd2tid_approx}
\alias{ymd2tid_approx}
\title{Transform a date into an time_id index}
\usage{
ymd2tid_approx(ymd, origin, period, yearly)
}
\arguments{
\item{ymd}{An int. A YYYYMMDD date}

\item{origin}{An int. A YYYYMMDD date. The day when the time_id == 0}

\item{period}{An int. The number of days between observations}

\item{yearly}{A boolean. Do the dates yearly match January the 1st?}
}
\value{
An integer. The time_id matching ymd or 0 (the first id) is ymd doesn't match
}
\description{
Transform a date into an time_id index
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
