% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ymd2ymd}
\alias{ymd2ymd}
\title{Split a date into parts}
\usage{
ymd2ymd(ymd)
}
\arguments{
\item{ymd}{An int YYYYMMDD}
}
\value{
An integer vector (year, month, day)
}
\description{
Split a date into its numeric parts
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
