% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{date2ydoy}
\alias{date2ydoy}
\title{Date to year-day-of-the-year}
\usage{
date2ydoy(dateAsText)
}
\arguments{
\item{dateAsText}{Date represented as a text string}
}
\value{
Character representing a date as day-of-the-year (YYYYDOY)
}
\description{
Transforms a date into the year-day_of_the_year date (YYYYDOY).
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
