% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{getxyMatrix}
\alias{getxyMatrix}
\title{Get the coordinates of the center of the given pixel}
\usage{
getxyMatrix(colrowid.Matrix, pixelSize)
}
\arguments{
\item{colrowid.Matrix}{A numeric matrix with 2 columns: col_id and row_id}

\item{pixelSize}{Pixel size in meters}
}
\value{
A 2-column matrix
}
\description{
Return the coords (MODIS synusoidal SR-ORG:6974) of the center of the given pixel
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
