% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{time_id2ydoy}
\alias{time_id2ydoy}
\title{Transform a time_id into year-day_of_the_year}
\usage{
time_id2ydoy(time_id, period)
}
\arguments{
\item{time_id}{A time id}

\item{period}{days between images (MOD09Q1 is 8, MOD13Q1 is 16)}
}
\value{
A number
}
\description{
DEPRECATED. Transform a time_id into year-day_of_the_year
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
