% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidbutil.R
\name{wgs84gmpi}
\alias{wgs84gmpi}
\title{Get the col_id & row_id in WGS84 coordinates}
\usage{
wgs84gmpi(lonlat.mat, pixelSize)
}
\arguments{
\item{lonlat.mat}{A numeric matrix with 2 columns: lon and lat on WGS84}

\item{pixelSize}{Pixel size in meters}
}
\value{
A 2-column matrix (col_id and row_id)
}
\description{
Calculate the col_id & row_id corresponding to the given WGS84 coordinates
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
